Refactored measure module
- Extended and updated the Measure base class
- Retired RadiancemeterArrayMeasure classes
- Introduced RadiancemeterMeasure
- Updated API reference
- Revived the SPP splitting feature
- Reqs: Moved tinydb to docs requirements
- Reqs: Updated locked dependencies
