+++ eradiate/solvers/onedim/_scene.py
@@ -6,7 +6,7 @@ from ..._attrs import documented, get_doc, parse_docs
 from ...scenes.atmosphere import Atmosphere, AtmosphereFactory, HomogeneousAtmosphere
 from ...scenes.core import KernelDict
 from ...scenes.integrators import Integrator, IntegratorFactory, VolPathIntegrator
-from ...scenes.measure import DistantMeasure, TargetOriginPoint, TargetOriginSphere
+from ...scenes.measure._distant import DistantMeasure, TargetOriginPoint, TargetOriginSphere
 from ...scenes.surface import LambertianSurface, Surface, SurfaceFactory
