+++ requirements/docs.in
@@ -7,3 +7,4 @@ sphinxcontrib-bibtex>=2.0
 sphinx-copybutton
 sphinx-gallery
 sphinx-panels
+tinydb
