+++ requirements/docs.txt
@@ -21,6 +21,10 @@ babel==2.9.0
     # via
     #   -c requirements/dev.txt
     #   sphinx
+beautifulsoup4==4.9.3
+    # via
+    #   -c requirements/dev.txt
+    #   pydata-sphinx-theme
 cerberus==1.3.2
     # via
     #   -c requirements/dev.txt
@@ -141,7 +145,7 @@ pybtex==0.24.0
     #   -c requirements/dev.txt
     #   pybtex-docutils
     #   sphinxcontrib-bibtex
-pydata-sphinx-theme==0.4.3
+pydata-sphinx-theme==0.5.0
     # via
     #   -c requirements/dev.txt
     #   -r requirements/docs.in
@@ -197,6 +201,10 @@ snowballstemmer==2.1.0
     # via
     #   -c requirements/dev.txt
     #   sphinx
+soupsieve==2.2
+    # via
+    #   -c requirements/dev.txt
+    #   beautifulsoup4
 sphinx-copybutton==0.3.1
     # via
     #   -c requirements/dev.txt
@@ -249,7 +257,7 @@ sphinxcontrib-serializinghtml==1.1.4
 tinydb==4.4.0
     # via
     #   -c requirements/dev.txt
-    #   -r requirements/main.in
+    #   -r requirements/docs.in
 tqdm==4.59.0
     # via
     #   -c requirements/dev.txt
