+++ setup.cfg
@@ -25,7 +25,6 @@ install_requires =
     pinttrs
     ruamel.yaml
     scipy
-    tinydb
     tqdm
     xarray
 
@@ -66,6 +65,7 @@ docs =
     sphinx-copybutton
     sphinx-gallery
     sphinx-panels
+    tinydb
 
 [options.entry_points]
 console_scripts =
